requirements.txt
absl-py==2.1.0
albucore==0.0.17
albumentations==1.4.17
annotated-types==0.7.0
astunparse==1.6.3
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dnspython==2.6.1
easydict==1.13
eval_type_backport==0.2.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnx==1.16.2
onnxruntime-gpu
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prettytable==3.11.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchtower==3.3.1
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
torch
fastapi
uvicorn
torchvision